refactor(server-actix): better structure
